wheelhouse: Updates to support deployment on Python 3.7
When deploying on a system with Python 3.7 only the currently
pinned pip version will be overridden as the pin set does not
include a version of pip with Python 3.7 support.

It appears the pinning was there in the first place to avoid a
bootstrapping issue like what can be seen in LP: #1824112.

By adding ``wheel`` to the wheelhouse we avoid this issue.

Unpin pip and setuptools.